refactor: FilaCheck fixes, PHPStan cleanup, layout component conversion, new enums
FilaCheck (27/8/155 → 30/5/70):
- Add table filters to 8 resource tables (status, is_active, reward_type, is_blocked)
- Add searchable columns to CapacityLimits and NotesRelationManager
- Migrate 24 deprecated callTableAction() to TestAction::make()->table()
- Add HasLabel to 6 enums (PlatformSenderType, LoyaltyPointType, SupportTicket*)
- Add filter indicators to WaitlistEntries, Expenses, Incomes
- Add canAccess() to AnnouncementBanner, CustomDomain, ManagePageContent
- Create custom Filament theme for admin panel
- Fix EmailCampaign status column enum type hints

PHPStan (600 → 286):
- Fix Product model docblock for UseEloquentBuilder attribute
- Add missing imports (StockAdjustmentType, Number, Builder)
- Fix return types for Number::format/currency casts
- Fix real bugs: invoice price→unit_price, catering photo→photo_path,
  blog feed author_name→storeName, remove dead category image code
- Add type annotations for generics and array params

Layout conversion:
- Convert storefront layout from @extends to <x-layouts.storefront> component
- Create Storefront component class with typed props
- Convert 14 storefront views to Blade component syntax
- Remove StorefrontComposer (component handles its own data)

New enums:
- AnnouncementType (info/warning/success/maintenance) cast on PlatformAnnouncement
- SupportReplyAuthorType (admin/tenant) cast on SupportReply

Refactoring plan completion:
- Consolidate FinancialReport to delegate to FinancialCalculator
- Replace RevenueChartWidget inline query with RevenueQuery::dailyBreakdown()
- BlogPostsTable filter uses BlogPostCategory enum directly

+++ app/Filament/Pages/PricingEngine.php
@@ -48,7 +48,7 @@ class PricingEngine extends Page
 
     public float $ingredientCost = 0;
 
-    public float $prepTimeMinutes = 0;
+    public int $prepTimeMinutes = 0;
 
     public float $hourlyLaborRate = 15;
 
@@ -92,7 +92,7 @@ class PricingEngine extends Page
         $recipe = $product->recipes->first();
         if ($recipe) {
             $this->ingredientCost = (float) ($recipe->cost ?? $product->cost ?? 0);
-            $this->prepTimeMinutes = (float) ($recipe->prep_time_minutes ?? 0);
+            $this->prepTimeMinutes = (int) ($recipe->prep_time_minutes ?? 0);
         } else {
             $this->ingredientCost = (float) ($product->cost ?? 0);
             $this->prepTimeMinutes = 0;
